pyproject.toml
[project]
name = "ai-chat"
version = "0.1.0"
description = "Django + DRF chat with Gemini and Vite frontend"
readme = "IMPLEMENTATION_PLAN.md"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0",
  "djangorestframework>=3.14",
  "python-dotenv>=1.0",
  "google-generativeai>=0.8",
  "pytest>=8.0",
  "pytest-django>=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ai_chat.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-ra"

